splitted webpack config in separate environment files for better understanding
also updated config files to typescript
