+++ package.json
@@ -14,15 +14,21 @@
         "styled-components": "^3.4.10"
     },
     "devDependencies": {
+        "@types/node": "^10.12.1",
         "@types/react": "^16.4.18",
         "@types/react-dom": "^16.0.9",
         "@types/react-redux": "^6.0.9",
         "@types/recompose": "^0.26.5",
+        "@types/uglifyjs-webpack-plugin": "^1.1.0",
+        "@types/webpack": "^4.4.17",
+        "@types/webpack-dev-server": "^3.1.1",
+        "@types/webpack-merge": "^4.1.3",
         "css-loader": "^1.0.1",
         "file-loader": "^1.1.11",
         "prettier": "^1.14.3",
         "style-loader": "^0.21.0",
         "ts-loader": "^4.5.0",
+        "ts-node": "^7.0.1",
         "tslint": "^5.11.0",
         "tslint-config-prettier": "^1.15.0",
         "tslint-react": "^3.6.0",
@@ -31,11 +37,12 @@
         "uglifyjs-webpack-plugin": "^1.3.0",
         "webpack": "^4.23.1",
         "webpack-cli": "^3.1.2",
-        "webpack-dev-server": "^3.1.10"
+        "webpack-dev-server": "^3.1.10",
+        "webpack-merge": "^4.1.4"
     },
     "scripts": {
-        "start": "webpack-dev-server --config ./webpack.config.js --host 0.0.0.0 --mode development",
-        "build": "webpack --config ./webpack.config.js --colors --progress --mode production"
+        "start": "webpack-dev-server --open --config ./webpack.dev.ts",
+        "build": "webpack --config ./webpack.prod.ts"
     },
     "author": "Alexander Kaufmann <ak@vivid-planet.com>",
     "license": "ISC"
